move api.impl to /common/
